chore: added tests , split files
