+++ crates/libtau/src/lib.rs
@@ -4,6 +4,7 @@ pub mod executor;
 pub mod metrics;
 pub mod model;
 pub mod ql;
+pub(crate) mod query;
 pub mod storage;
 pub mod users;
 pub mod value;
